+++ src/apis/auth/discord.py
@@ -175,6 +175,7 @@ async def get_callback(request: Request, response: Response, code: Optional[str]
             return {"error": ml.tr(request, "unknown_error")}
 
     except Exception as exc:
-        await tracebackHandler(request, exc, traceback.format_exc())
+        if not isinstance(exc, RateLimitException):
+            await tracebackHandler(request, exc, traceback.format_exc())
         response.status_code = 400
         return {"error": ml.tr(request, "unknown_error")}